+++ prisma/migrations/20241231111718_user_tracking_enums/migration.sql
@@ -1,0 +1,2 @@
+-- AlterTable
+ALTER TABLE `UserTrackingList` ADD COLUMN `type` ENUM('FRIENDS', 'ACHTUNGS', 'OTHER') NOT NULL DEFAULT 'OTHER';