+++ src/components/views/header/ViewHeaderSettings.vue
@@ -3,11 +3,21 @@
         v-model="model"
         absolute
         class="settings"
-        :sections="[
-            { title: 'VATSIM Account', key: 'account' },
-            { title: 'Navigraph Account', key: 'navigraph' },
-            { title: 'Preferences', key: 'follow' },
-        ]"
+        max-height="calc(100dvh - 56px - 92px - 20px)"
+        :tabs="{
+            main: {
+                title: 'Settings',
+                sections: [
+                    { title: 'VATSIM Account', key: 'account' },
+                    { title: 'Navigraph Account', key: 'navigraph' },
+                    { title: 'Preferences', key: 'follow' },
+                ],
+            },
+            favorite: {
+                title: 'Favorite Lists',
+                sections: listsSections,
+            },
+        }"
     >
         <template #title>
             Settings
@@ -108,6 +118,22 @@
                 </common-toggle>
             </div>
         </template>
+        <template #new-list>
+            <view-header-list
+                :list="newList"
+                @add="addList"
+            />
+        </template>
+        <template
+            v-for="(list, index) in store.lists"
+            :key="list.id"
+            #[list.id]
+        >
+            <view-header-list
+                :last="index === store.lists.length - 1"
+                :list
+            />
+        </template>
         <template #navigraph>
             <div class="settings__block settings__block--long-gap">
                 <common-button
@@ -144,9 +170,8 @@
                     </common-button>
                 </div>
                 <div class="settings__description">
-                    Users with linked Navigraph and Unlimited/Data subscription will receive latest AIRAC for
-                    gates, waypoints and all
-                    other Navigraph-related stuff
+                    Users with linked Navigraph and Unlimited/Data subscription receive latest AIRAC for
+                    gates and runways. Navigraph Ultimate members also get Airport Layouts feature.
                 </div>
             </div>
         </template>
@@ -155,6 +180,7 @@
 
 <script setup lang="ts">
 import CommonInfoPopup from '~/components/common/popup/CommonInfoPopup.vue';
+import type { InfoPopupSection } from '~/components/common/popup/CommonInfoPopup.vue';
 import CommonToggle from '~/components/common/basic/CommonToggle.vue';
 import CommonButton from '~/components/common/basic/CommonButton.vue';
 import CommonInfoBlock from '~/components/common/blocks/CommonInfoBlock.vue';
@@ -163,6 +189,9 @@ import CommonInputText from '~/components/common/basic/CommonInputText.vue';
 import { useStore } from '~/store';
 import { defu } from 'defu';
 import type { UserSettings } from '~/utils/backend/user';
+import ViewHeaderList from '~/components/views/header/ViewHeaderList.vue';
+import type { UserListLive } from '~/utils/backend/lists';
+import { MAX_USER_LISTS } from '~/utils/shared';
 
 const model = defineModel({ type: Boolean, required: true });
 const store = useStore();
@@ -172,6 +201,34 @@ const settings = reactive(defu<UserSettings, [UserSettings]>(store.user?.setting
     autoZoom: false,
 }));
 
+const listsSections = computed<InfoPopupSection[]>(() => {
+    const sections = store.lists.map(x => ({
+        title: x.name,
+        key: x.id.toString(),
+        list: x,
+        collapsible: true,
+    })) satisfies InfoPopupSection[] as InfoPopupSection[];
+
+    if (sections.length < MAX_USER_LISTS) {
+        sections.unshift({
+            title: 'New List',
+            key: 'new-list',
+            collapsible: true,
+            collapsedDefault: true,
+            collapsedDefaultOnce: true,
+        });
+    }
+
+    return sections;
+});
+
+useClickOutside({
+    element: () => document.querySelector('.info-popup'),
+    callback: e => {
+        model.value = false;
+    },
+});
+
 watch(settings, () => {
     $fetch('/api/user/settings', {
         method: 'POST',
@@ -179,6 +236,27 @@ watch(settings, () => {
     });
     store.user!.settings = settings;
 });
+
+const newList = reactive<UserListLive>({
+    id: -1,
+    name: '',
+    type: 'OTHER',
+    color: 'info500',
+    users: [],
+    showInMenu: false,
+});
+
+async function addList() {
+    await addUserList(newList);
+
+    Object.assign(newList, {
+        id: -1,
+        name: '',
+        type: 'OTHER',
+        color: 'info500',
+        users: [],
+    });
+}
 </script>
 
 <style scoped lang="scss">