+++ src/store/index.ts
@@ -10,6 +10,7 @@ import { useMapStore } from '~/store/map';
 import type { Coordinate } from 'ol/coordinate';
 import type { UserMapPreset, UserMapSettings } from '~/utils/backend/map-settings';
 import type { TurnsBulkReturn } from '~/server/api/data/vatsim/pilot/turns';
+import type { UserListLive, UserListLiveUser } from '~/utils/backend/lists';
 
 export interface SiteConfig {
     hideSectors?: boolean;
@@ -51,6 +52,7 @@ export const useStore = defineStore('index', {
 
         featuredAirportsOpen: false,
         featuredVisibleOnly: false,
+        menuFriendsOpen: false,
 
         updateRequired: false,
         isTabVisible: false,
@@ -84,6 +86,87 @@ export const useStore = defineStore('index', {
             }
 
             return this.theme;
+        },
+        lists(): UserListLive[] {
+            if (!this.user) return [];
+
+            const lists = this.user.lists.slice(0);
+            const listsUsers = new Set(lists.flatMap(x => x.users.map(x => x.cid)));
+            const foundUsers: Record<number, Pick<UserListLiveUser, 'type' | 'data'>> = {};
+
+            const dataStore = useDataStore();
+
+            if (!lists.some(x => x.type === 'FRIENDS')) {
+                lists.unshift({
+                    id: 0,
+                    name: 'Friends',
+                    color: 'success300',
+                    type: 'FRIENDS',
+                    showInMenu: true,
+                    users: [],
+                });
+            }
+
+            if (listsUsers.size) {
+                for (const pilot of dataStore.vatsim.data.pilots.value) {
+                    if (listsUsers.has(pilot.cid)) {
+                        foundUsers[pilot.cid] = {
+                            type: 'pilot',
+                            data: pilot,
+                        };
+                    }
+                }
+
+                for (const atc of dataStore.vatsim.data.locals.value) {
+                    if (atc.atc.isATIS) continue;
+                    if (listsUsers.has(atc.atc.cid)) {
+                        foundUsers[atc.atc.cid] = {
+                            type: 'atc',
+                            data: atc.atc,
+                        };
+                    }
+                }
+
+                for (const atc of dataStore.vatsim.data.firs.value) {
+                    if (atc.controller.isATIS) continue;
+                    if (listsUsers.has(atc.controller.cid)) {
+                        foundUsers[atc.controller.cid] = {
+                            type: atc.controller.rating === 1 ? 'sup' : 'atc',
+                            data: atc.controller,
+                        };
+                    }
+                }
+
+                for (const prefile of dataStore.vatsim.data.prefiles.value) {
+                    if (listsUsers.has(prefile.cid)) {
+                        foundUsers[prefile.cid] = {
+                            type: 'prefile',
+                            data: prefile,
+                        };
+                    }
+                }
+            }
+
+            return lists.map(list => ({
+                ...list,
+                users: list.users.map(user => ({
+                    ...user,
+                    ...foundUsers[user.cid] ?? {
+                        type: 'offline',
+                        data: undefined,
+                    },
+                } as UserListLiveUser)).sort((a, b) => {
+                    const aOnline = a.type !== 'offline';
+                    const bOnline = b.type !== 'offline';
+
+                    if (bOnline && !aOnline) return 1;
+                    if (!bOnline && aOnline) return -1;
+                    return 0;
+                }),
+            }));
+        },
+        friends(): UserListLiveUser[] {
+            return this.lists.filter(x => x.showInMenu).flatMap(x => x.users.filter(x => x.type !== 'offline'));
         },
     },
     actions: {
@@ -140,6 +223,9 @@ export const useStore = defineStore('index', {
             finally {
                 this.dataInProgress = false;
             }
+        },
+        async refreshUser() {
+            this.user = await $fetch<FullUser>('/api/user');
         },
     },
 });